Apply separate weights for unet and text encoder
This Pull Request makes it possible to apply separate weights to the unet and text encoder, which allows for finer adjustments after training. This should be pretty useful since training both the unet and tenc perfectly is pretty hard.

I tested that the resulting images with tenc_weight == unet_weight are equal to the resulting images from before this change with the same weight.